requirements.txt
paho-mqtt==2.1.0
google-cloud-bigquery==3.26.0